Initial commit. Put views in separate folder
