+++ MaxCard/views/LoadingScreen.js
